+++ CHANGELOG.md
@@ -3,7 +3,17 @@
 
 ## Improvements
 
-- WIP
+Addition of the Asset manager end-points:
+
+- Get an asset family
+- Get a list of asset families
+- Upsert an asset family
+- Get an asset family attribute
+- Get a list of asset family attributes
+- Upsert an asset family attribute
+- Get an asset family attribute option
+- Get a list of asset family attribute options
+- Upsert an asset family attribute option
 
 # 5.0.0